pyproject.toml
[project]
name = "quantitative-trading"
version = "0.1.0"
description = "Rule One value-investing agent (LLM-driven qualitative + code-driven quantitative) with point-in-time historical backtesting on SEC EDGAR + yfinance."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Michele Scalzotto" }]

dependencies = [
    "anthropic>=0.40.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "pyarrow>=15.0.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "tenacity>=8.2.0",
    "tqdm>=4.66.0",
    "yfinance>=0.2.40",
]

[project.optional-dependencies]
notebook = [
    "ipykernel>=6.29.0",
    "jupyter>=1.0.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/quantitative_trading"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM", "RET", "PL"]
ignore = ["PLR0913", "PLR2004"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
